+++ bc/alerts/wagtail_hooks.py
@@ -6,7 +6,7 @@ from bc.alerts.models import Alert
 class AlertModelAdmin(ModelAdmin):
     model = Alert
     menu_icon = "warning"
-    list_display = ("title", "page")
+    list_display = ("title", "alert_level", "page")
 
 
 modeladmin_register(AlertModelAdmin)