+++ bc/recruitment/management/commands/send_job_alerts.py
@@ -5,6 +5,8 @@ from django.http import QueryDict
 from django.template.loader import render_to_string
 from django.utils.timezone import now
 
+from django_gov_notify.message import NotifyEmailMessage
+
 from bc.recruitment.models import (
     JobAlertNotificationTask,
     JobAlertSubscription,
@@ -12,7 +14,6 @@ from bc.recruitment.models import (
     TalentLinkJob,
 )
 from bc.recruitment.utils import get_job_search_results
-from bc.utils.email import NotifyEmailMessage
 
 
 class Command(BaseCommand):