+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Release History
 
+## Unreleased
+
+- 40: Feedback forms and reports
+- Accessibility improvements
+- Add multiple severity levels to sitewide alert
+- Fix: StreamField typo causing RecursionError in Wagtail 2.13
+- Upgrade Wagtail to version 2.13
+- Use the newly released django-gov-notify library for GOV.UK Notify integration
+
 ## 36.1 (2021-05-25)
 
 Compare: https://git.torchbox.com/buckinghamshire-council/bc/compare/36.0...36.1